Populate dispatch info in dispatch message.

+++ ORN-Ops/ORN-Ops/TeamDetailTableViewController.m
@@ -234,7 +234,7 @@
 	
 	// Load dispatch fields
 	self.isActiveSwitch.on = self.team.isActive.boolValue;
-	self.rideCountLabel.text = [NSString stringWithFormat:@"%d", (int)self.team.ridesAssigned.count];
+	self.rideCountLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)self.team.ridesAssigned.count];
 	self.durationLabel.text = [NSString stringWithFormat:@"%.0f min", self.team.assignedDuration / (NSTimeInterval)SECONDS_PER_MINUTE];
 	self.distanceLabel.text = [NSString stringWithFormat:@"%.1f km", self.team.assignedDistance / (CLLocationDistance)METERS_PER_KILOMETER];
 	self.donationsLabel.text =  [self.currencyNumberFormatter stringFromNumber:self.team.assignedDonations];
